backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# OpenAI
openai==1.12.0

# Vector Database
qdrant-client==1.7.3

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS
fastapi-cors==0.0.6

# Markdown processing
markdown==3.5.2
beautifulsoup4==4.12.3

# Rate limiting
slowapi==0.1.9
